Revert "修复failover失败问题"
This reverts commit 7c9ff1aa6706e86cc1aed08159b5e25986df39dd.

Revert "修复遗漏opengauss宏"

This reverts commit 06492275dea116b60e34bf2d44a6d171d43e7ae5.

Revert "屏蔽lsnd_wait"

This reverts commit 50323d55c445682357902b125091f2419bf36b11.

Revert "增加htap delta相关函数"

This reverts commit 343f71b38c7116e3cac14691dbfd88131211d238.

Revert "fix sys dependency"

This reverts commit 4d0a2ab6f97f2b3dda1c0d46991993cb30971453.

Revert "分布式锁加锁过程中，能及时感知kill session及死锁等session 状态"

This reverts commit a5a5f4db3c3e4b51aabcb41ef0c1388561d81bc6.

Revert "DMS代码白盒检视修改"

This reverts commit 2f11a440bd7e1e88a83854c4c79af65f4526cec0.

Revert "pagehash val calc by group"

This reverts commit dd005f743d11c1638b72530da84aa1ed9e6059f7.

Revert "sync: 1. bcast code refactor 2. trc bugfix"

This reverts commit e0912046441ea219d559360955a935abef7f07b1.

Revert "fix compile errors"

This reverts commit c74719da6a9254c223f4562f181af005ff0b27a6.

Revert "优化结构体dms_time_consume_t的空间占用"

This reverts commit 773ca4b8e0fc9949c3abde68e940cc078c07dd13.

Revert "DMS等待事件直接登记到DB侧"

This reverts commit 972b5e93794390bf405bbb3547ae05a38f8e74b3.

Revert "优化圈复杂度"

This reverts commit 217ae7961b3ffe4efc2d0015c058ffeac31ce8c3.

Revert "兼容性相关的数据结构/变量/宏定义/函数定义出来"

This reverts commit fe54872e9e5c30575a3cc15e837ec0049e7fd527.

Revert "bugfix"

This reverts commit 17fb7467b401e03c1107814d789943a48a49668d.

Revert "reform各节点属性放到ctx中"

This reverts commit 85cc102e54f0502ac43872d6f86c4e6192c90cf5.

Revert "新增reform步骤-兼容性适配"

This reverts commit 8e31dbd80d565d65a85b4afb963b5323ce470f1f.

Revert "优化日志打印"

This reverts commit b9ff9fdfa1741077f1f3fdf642dba8c655240f28.

Revert "codecheck"

This reverts commit a7361a1587cf0b112669e583c91df2f7d9e19c24.

Revert "worker视图新增字段"

This reverts commit 3584afea28d38ceb0e17ad1464a8c632cb1e83ae.

Revert "降低圈复杂度"

This reverts commit 21209bcec89638646e21842081640a1604326cdc.

Revert "修改edp页面的last_edp_lsn>已刷盘的lsn，导致edp清理失败的问题"

This reverts commit 4b0f86b29da20bc4004d291ce6186e2caad076d7.

Revert "codecheck"

This reverts commit cba8bf364a7968f50deaf300b50511320cbb6124.

Revert "dms新增创建drc的消息"

This reverts commit cfa8cdaf80a4fb1b24b99412d40238adfb321ee6.

Revert "大文件拆分"

This reverts commit ed6866e2f70cb58d012550a6f4c2078f4be31e0e.

Revert "分布式锁DRC及本地lock res支持回收功能"

This reverts commit c201b42fe895b1b85d0193e620ac8d9db6e42936.

Revert "新增dms相关接口"

This reverts commit 3a1f7038b45efe9962266d4f9999b86b8662db78.

Revert "支持DRC动态迁移"

This reverts commit e77736ffedf7393c60d88bd2dbd447db254325bd.

Revert "回复ack消息时，seq=0问题修复"

This reverts commit aca2c603de0a7638fd266d1779aeb8c115153840.

Revert "edp页面误清理导致脏页队列断链问题修复"

This reverts commit d19cc20b8c9562cd5a7c93b5d022ab17cd62dab2.

Revert "DRC管理结构优化补充"

This reverts commit ec16019f369e1c009882e89faa33480a079e4b72.

Revert "分布式锁的失效和cancel时序存在问题，导致drc上copy_inst信息误清理"

This reverts commit 1955d14fc8ae46f28c53137071a8c465602bdde7.

Revert "【代码检视】DB层XID定义为8字节，DMS层XID定义12字节，存在越界访问"

This reverts commit 7f62f6c693d6071e01563a3934a1305906d3dd4f.

Revert "fix"

This reverts commit cd2a45f1b5e1d05725e08957febf711e593c0c29.

Revert "适配DSS修改log_type_t里的枚举名称"

This reverts commit f3e8404a276804fdd4859bb98f004e0cafc02287.

Revert "add support for expansion"

This reverts commit 1f1d23606a8e310067688750beeb986f12d17156.

Revert "解决reform卡住问题"

This reverts commit 4629adea6fc65465df9f5773849351ad9df815d5.

Revert "reform消息和业务消息分组"

This reverts commit fd7b819a5f1bdde5c88bdc68a2375853801c2cb3.

Revert "fix"

This reverts commit 1d42b7147ac640ec13f1edf70460b1c91642dbcc.

Revert "删除mfc"

This reverts commit c0348afb84efa151474863fd4138c8fd3b0225cf.

Revert "add event waiting interface;"

This reverts commit 899e732ea3e371762830372be16709de2069b9ba.

Revert "fix"

This reverts commit f8e98450a75e3e955dd2acf1fdc6b0f7c2192378.

Revert "DMS reform 事件通知"

This reverts commit 34aa42d7602ae6cb9a16b5ac5d700325a0af5d84.

+++ src/drc/drc_page.h
@@ -37,8 +37,7 @@ extern "C" {
 
 static inline bool32 if_cvt_need_confirm(drc_cvt_item_t *converting)
 {
-    if (((int64)(g_timer()->monotonic_now - converting->begin_time) / (int64)MICROSECS_PER_MILLISEC)
-        > DMS_CVT_EXPIRE_TIME) {
+    if (((g_timer()->monotonic_now - converting->begin_time) / (int64)MICROSECS_PER_MILLISEC) > DMS_CVT_EXPIRE_TIME) {
         return CM_TRUE;
     }
     return CM_FALSE;
